몬스터 세부 조정 / 플레이어 스탯up관련 코드 - paperItem코드 수정 / 중간백업

+++ Assets/Game/Scripts/Monster/MonsterStatus.cs
@@ -5,8 +5,8 @@ using UnityEngine;
 [System.Serializable]
 public class MonsterStatus 
 {
-    public int hp;
-    public int maxHp;
+    public float hp;
+    public float maxHp;
 
     public float shortMonsterRange; //플레이어로 이동하기 위한 거리
     public float longMonsterRange;
